requirements.txt
h5py==3.4.0
hydra-core==1.0.7
ninja==1.10.2
numexpr==2.8.4
openpyxl==3.0.10
numpy==1.26.4
pandas==1.4.2
scikit-learn==0.24.1
transformers==4.37.2
scipy
tokenizers==0.15.2
torch==1.13.0
torchvision==0.14.0
netcal==1.3.5
matplotlib>=2.2.2
datasets==2.4.0
tqdm
numpy
open_clip_torch
git+https://github.com/openai/CLIP.git
faiss-gpu
spacy
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
